+++ quick_test/quick_run.py
@@ -1,17 +1,16 @@
 from typing import List
 
-def test_run(nums: List[int], target: int) -> List[int]:
-    hash_map = {}
-    for i, v in enumerate(nums):
-        search = target - v
-        if search in hash_map:
-            return [hash_map[search], i]
-        
-        hash_map[v] = i
+def test_run(strs: List[str]) -> str:
+    result = ''
+    for letter_set in zip(*strs):
+        if len(set(letter_set)) == 1:
+            result += letter_set[0]
+        else:
+            break
 
-    return
+    return result
 
 
 if __name__ == '__main__':
-    result = test_run([2,7,11,15], 9)
+    result = test_run(["flower","flow","flight"])
     print(f"result: {result}")
